build: 공통으로 사용되는 파일 src/include로 이동, 서식 수정
- 라이브러리 헤더 주석 추가
- 스페이스 탭으로 변경
- 설정 파일 확장자 .conf로 변경
